pyproject.toml
[tool.poetry]
name = "dj-eshop"
version = "0.1.0"
description = ""
authors = ["Saken Myrzabekov <myrzabekovsaken89@gmail.com>"]
readme = "README.md"
packages = [{include = "dj_eshop"}]

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
markdown = "^3.5"
django-filter = "^23.3"
django-cors-headers = "^4.2.0"
djoser = "^2.2.0"
djangorestframework-simplejwt = "^5.3.0"
pillow = "^10.0.1"
razorpay = "^1.4.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
